Move class-level classes to package-level from Resources
