+++ src/service/resources-impl.js
@@ -14,33 +14,29 @@
  * limitations under the License.
  */
 
+import {FiniteStateMachine} from '../finite-state-machine';
 import {FocusHistory} from '../focus-history';
 import {Pass} from '../pass';
+import {Resource, ResourceState} from './resource';
+import {TaskQueue} from './task-queue';
+import {VisibilityState} from '../visibility-state';
+import {checkAndFix as ieMediaCheckAndFix} from './ie-media-bug';
 import {closest, hasNextNodeInDocumentOrder} from '../dom';
 import {onDocumentReady} from '../document-ready';
-import {
-  expandLayoutRect,
-  layoutRectLtwh,
-  layoutRectsOverlap,
-} from '../layout-rect';
+import {expandLayoutRect} from '../layout-rect';
 import {getService} from '../service';
 import {inputFor} from '../input';
-import {dev} from '../log';
-import {reportError} from '../error';
-import {timer} from '../timer';
 import {installFramerateService} from './framerate-impl';
 import {installViewerService} from './viewer-impl';
 import {installViewportService} from './viewport-impl';
 import {installVsyncService} from './vsync-impl';
-import {platformFor} from '../platform';
-import {FiniteStateMachine} from '../finite-state-machine';
 import {isArray} from '../types';
-import {VisibilityState} from '../visibility-state';
+import {dev} from '../log';
+import {reportError} from '../error';
+import {timer} from '../timer';
 
 
 const TAG_ = 'Resources';
-const RESOURCE_PROP_ = '__AMP__RESOURCE';
-const OWNER_PROP_ = '__AMP__OWNER';
 const LAYOUT_TASK_ID_ = 'L';
 const LAYOUT_TASK_OFFSET_ = 0;
 const PRELOAD_TASK_ID_ = 'P';
@@ -65,15 +61,15 @@ const FOUR_FRAME_DELAY_ = 70;
 let ChangeSizeRequestDef;
 
 export class Resources {
+  /**
+   * @param {!Window} window
+   */
   constructor(window) {
     /** @const {!Window} */
     this.win = window;
 
     /** @const @private {!./viewer-impl.Viewer} */
     this.viewer_ = installViewerService(window);
-
-    /** @const @private {!../platform.Platform} */
-    this.platform_ = platformFor(window);
 
     /** @private {boolean} */
     this.isRuntimeOn_ = this.viewer_.isRuntimeOn();
@@ -124,11 +120,11 @@ export class Resources {
     /** @const {!Pass} */
     this.pass_ = new Pass(() => this.doPass_());
 
-    /** @const {!TaskQueue_} */
-    this.exec_ = new TaskQueue_();
+    /** @const {!TaskQueue} */
+    this.exec_ = new TaskQueue();
 
-    /** @const {!TaskQueue_} */
-    this.queue_ = new TaskQueue_();
+    /** @const {!TaskQueue} */
+    this.queue_ = new TaskQueue();
 
    /**
     * @private {!Array<!ChangeSizeRequestDef>}
@@ -201,9 +197,14 @@ export class Resources {
       this.documentReady_ = true;
       this.buildReadyResources_();
       this.pendingBuildResources_ = null;
-      if (this.platform_.isIe()) {
-        this.fixMediaIe_(this.win);
+      const fixPromise = ieMediaCheckAndFix(this.win);
+      if (fixPromise) {
+        fixPromise.then(() => {
+          this.relayoutAll_ = true;
+          this.schedulePass();
+        });
       } else {
+        // No promise means that there's no problem.
         this.relayoutAll_ = true;
       }
       this.schedulePass();
@@ -214,57 +215,20 @@ export class Resources {
   }
 
   /**
-   * An ugly fix for IE's problem with `matchMedia` API, where media queries
-   * are evaluated incorrectly. See #2577 for more details.
-   * @param {!Window} win
-   * @private
-   */
-  fixMediaIe_(win) {
-    if (!this.platform_.isIe() || this.matchMediaIeQuite_(win)) {
-      this.relayoutAll_ = true;
-      return;
-    }
-
-    // Poll until the expression resolves correctly, but only up to a point.
-    const endTime = timer.now() + 2000;
-    const interval = win.setInterval(() => {
-      const now = timer.now();
-      const matches = this.matchMediaIeQuite_(win);
-      if (matches || now > endTime) {
-        win.clearInterval(interval);
-        this.relayoutAll_ = true;
-        this.schedulePass();
-        if (!matches) {
-          dev.error(TAG_, 'IE media never resolved');
-        }
-      }
-    }, 10);
-  }
-
-  /**
-   * @param {!Window} win
-   * @return {boolean}
-   * @private
-   */
-  matchMediaIeQuite_(win) {
-    const q = `(min-width: ${win./*OK*/innerWidth}px)` +
-        ` AND (max-width: ${win./*OK*/innerWidth}px)`;
-    try {
-      return win.matchMedia(q).matches;
-    } catch (e) {
-      dev.error(TAG_, 'IE matchMedia failed: ', e);
-      // Return `true` to avoid polling on a broken API.
-      return true;
-    }
-  }
-
-  /**
    * Returns a list of resources.
    * @return {!Array<!Resource>}
    * @export
    */
   get() {
     return this.resources_.slice(0);
+  }
+
+  /**
+   * Whether the runtime is currently on.
+   * @return {boolean}
+   */
+  isRuntimeOn() {
+    return this.isRuntimeOn_;
   }
 
   /**
@@ -333,12 +297,11 @@ export class Resources {
    * Returns the {@link Resource} instance corresponding to the specified AMP
    * Element. If no Resource is found, the exception is thrown.
    * @param {!AmpElement} element
-   * @return {?Resource}
+   * @return {!Resource}
    * @package
    */
   getResourceForElement(element) {
-    return dev.assert(/** @type {!Resource} */ (element[RESOURCE_PROP_]),
-        'Missing resource prop on %s', element);
+    return Resource.forElement(element);
   }
 
   /**
@@ -371,7 +334,6 @@ export class Resources {
     if (!element.id) {
       element.id = 'AMP_' + resource.getId();
     }
-    element[RESOURCE_PROP_] = resource;
     this.resources_.push(resource);
 
     if (this.isRuntimeOn_) {
@@ -431,7 +393,10 @@ export class Resources {
    * @package
    */
   remove(element) {
-    const resource = this.getResourceForElement(element);
+    const resource = Resource.forElementOptional(element);
+    if (!resource) {
+      return;
+    }
     const index = resource ? this.resources_.indexOf(resource) : -1;
     if (index != -1) {
       this.resources_.splice(index, 1);
@@ -447,7 +412,7 @@ export class Resources {
    * @package
    */
   upgraded(element) {
-    const resource = this.getResourceForElement(element);
+    const resource = Resource.forElement(element);
     if (this.isRuntimeOn_) {
       resource.build();
       this.schedulePass();
@@ -465,8 +430,7 @@ export class Resources {
    * @package
    */
   setOwner(element, owner) {
-    dev.assert(owner.contains(element), 'Owner must contain the element');
-    element[OWNER_PROP_] = owner;
+    Resource.setAsOwner(element, owner);
   }
 
   /**
@@ -480,7 +444,7 @@ export class Resources {
    */
   scheduleLayout(parentElement, subElements) {
     this.scheduleLayoutOrPreloadForSubresources_(
-        this.getResourceForElement(parentElement),
+        Resource.forElement(parentElement),
         /* layout */ true,
         elements_(subElements));
   }
@@ -492,7 +456,7 @@ export class Resources {
    * @param {!Element|!Array<!Element>} subElements
    */
   schedulePause(parentElement, subElements) {
-    const parentResource = this.getResourceForElement(parentElement);
+    const parentResource = Resource.forElement(parentElement);
     subElements = elements_(subElements);
 
     this.discoverResourcesForArray_(parentResource, subElements, resource => {
@@ -511,7 +475,7 @@ export class Resources {
    */
   schedulePreload(parentElement, subElements) {
     this.scheduleLayoutOrPreloadForSubresources_(
-        this.getResourceForElement(parentElement),
+        Resource.forElement(parentElement),
         /* layout */ false,
         elements_(subElements));
   }
@@ -527,7 +491,7 @@ export class Resources {
    */
   updateInViewport(parentElement, subElements, inLocalViewport) {
     this.updateInViewportForSubresources_(
-        this.getResourceForElement(parentElement),
+        Resource.forElement(parentElement),
         elements_(subElements),
         inLocalViewport);
   }
@@ -541,7 +505,7 @@ export class Resources {
    * @param {function()=} opt_callback A callback function.
    */
   changeSize(element, newHeight, newWidth, opt_callback) {
-    this.scheduleChangeSize_(this.getResourceForElement(element), newHeight,
+    this.scheduleChangeSize_(Resource.forElement(element), newHeight,
         newWidth, /* force */ true, opt_callback);
   }
 
@@ -564,7 +528,7 @@ export class Resources {
    * @protected
    */
   attemptChangeSize(element, newHeight, newWidth, opt_callback) {
-    this.scheduleChangeSize_(this.getResourceForElement(element), newHeight,
+    this.scheduleChangeSize_(Resource.forElement(element), newHeight,
         newWidth, /* force */ false, opt_callback);
   }
 
@@ -574,7 +538,7 @@ export class Resources {
    * @param {!Function} callback
    */
   deferMutate(element, callback) {
-    this.scheduleDeferredMutate_(this.getResourceForElement(element), callback);
+    this.scheduleDeferredMutate_(Resource.forElement(element), callback);
     this.schedulePassVsync();
   }
 
@@ -610,12 +574,12 @@ export class Resources {
 
         // Mark itself and children for re-measurement.
         if (element.classList.contains('-amp-element')) {
-          const r = this.getResourceForElement(element);
+          const r = Resource.forElement(element);
           r.requestMeasure();
         }
         const ampElements = element.getElementsByClassName('-amp-element');
         for (let i = 0; i < ampElements.length; i++) {
-          const r = this.getResourceForElement(ampElements[i]);
+          const r = Resource.forElement(ampElements[i]);
           r.requestMeasure();
         }
         if (relayoutTop != -1) {
@@ -860,11 +824,12 @@ export class Resources {
    * @private
    */
   checkPendingChangeSize_(element) {
-    const resourceElement = closest(element, el => el[RESOURCE_PROP_]);
+    const resourceElement = closest(element,
+        el => !!Resource.forElementOptional(el));
     if (!resourceElement) {
       return;
     }
-    const resource = this.getResourceForElement(resourceElement);
+    const resource = Resource.forElement(resourceElement);
     const pendingChangeSize = resource.getPendingChangeSize();
     if (pendingChangeSize !== undefined) {
       this.scheduleChangeSize_(resource, pendingChangeSize.height,
@@ -902,10 +867,10 @@ export class Resources {
     let remeasureCount = 0;
     for (let i = 0; i < this.resources_.length; i++) {
       const r = this.resources_[i];
-      if (r.getState() == ResourceState_.NOT_BUILT) {
+      if (r.getState() == ResourceState.NOT_BUILT) {
         r.build();
       }
-      if (relayoutAll || r.getState() == ResourceState_.NOT_LAID_OUT) {
+      if (relayoutAll || r.getState() == ResourceState.NOT_LAID_OUT) {
         r.applySizesAndMediaQuery();
         relayoutCount++;
       }
@@ -921,11 +886,11 @@ export class Resources {
             relayoutAll || relayoutTop != -1) {
       for (let i = 0; i < this.resources_.length; i++) {
         const r = this.resources_[i];
-        if (r.getState() == ResourceState_.NOT_BUILT || r.hasOwner()) {
+        if (r.getState() == ResourceState.NOT_BUILT || r.hasOwner()) {
           continue;
         }
         if (relayoutAll ||
-                r.getState() == ResourceState_.NOT_LAID_OUT ||
+                r.getState() == ResourceState.NOT_LAID_OUT ||
                 r.isMeasureRequested() ||
                 relayoutTop != -1 && r.getLayoutBox().bottom >= relayoutTop) {
           const wasDisplayed = r.isDisplayed();
@@ -971,7 +936,7 @@ export class Resources {
     if (loadRect) {
       for (let i = 0; i < this.resources_.length; i++) {
         const r = this.resources_[i];
-        if (r.getState() != ResourceState_.READY_FOR_LAYOUT || r.hasOwner()) {
+        if (r.getState() != ResourceState.READY_FOR_LAYOUT || r.hasOwner()) {
           continue;
         }
         // TODO(dvoytenko, #3434): Reimplement the use of `isFixed` with
@@ -1008,7 +973,7 @@ export class Resources {
       let idleScheduledCount = 0;
       for (let i = 0; i < this.resources_.length; i++) {
         const r = this.resources_[i];
-        if (r.getState() == ResourceState_.READY_FOR_LAYOUT &&
+        if (r.getState() == ResourceState.READY_FOR_LAYOUT &&
                 !r.hasOwner() && r.isDisplayed()) {
           dev.fine(TAG_, 'idle layout:', r.debugid);
           this.scheduleLayoutOrPreload_(r, /* layout */ false);
@@ -1069,6 +1034,9 @@ export class Resources {
         task.promise.then(this.taskComplete_.bind(this, task, true),
             this.taskComplete_.bind(this, task, false))
             .catch(reportError);
+        if (task.resource.getState() == ResourceState.LAYOUT_SCHEDULED) {
+          this.framerate_.collect(task.resource.element);
+        }
       }
 
       task = this.queue_.peek(scorer);
@@ -1267,7 +1235,7 @@ export class Resources {
    * @private
    */
   scheduleLayoutOrPreload_(resource, layout, opt_parentPriority) {
-    dev.assert(resource.getState() != ResourceState_.NOT_BUILT &&
+    dev.assert(resource.getState() != ResourceState.NOT_BUILT &&
         resource.isDisplayed(),
         'Not ready for layout: %s (%s)',
         resource.debugid, resource.getState());
@@ -1309,14 +1277,14 @@ export class Resources {
   scheduleLayoutOrPreloadForSubresources_(parentResource, layout, subElements) {
     const resources = [];
     this.discoverResourcesForArray_(parentResource, subElements, resource => {
-      if (resource.getState() != ResourceState_.NOT_BUILT) {
+      if (resource.getState() != ResourceState.NOT_BUILT) {
         resources.push(resource);
       }
     });
     if (resources.length > 0) {
       resources.forEach(resource => {
         resource.measure();
-        if (resource.getState() == ResourceState_.READY_FOR_LAYOUT &&
+        if (resource.getState() == ResourceState.READY_FOR_LAYOUT &&
                 resource.isDisplayed()) {
           this.scheduleLayoutOrPreload_(resource, layout,
               parentResource.getPriority());
@@ -1397,7 +1365,7 @@ export class Resources {
   discoverResourcesForElement_(element, callback) {
     // Breadth-first search.
     if (element.classList.contains('-amp-element')) {
-      callback(this.getResourceForElement(element));
+      callback(Resource.forElement(element));
       // Also schedule amp-element that is a placeholder for the element.
       const placeholder = element.getPlaceholder();
       if (placeholder) {
@@ -1417,7 +1385,7 @@ export class Resources {
         }
         if (!covered) {
           seen.push(ampElement);
-          callback(this.getResourceForElement(ampElement));
+          callback(Resource.forElement(ampElement));
         }
       }
     }
@@ -1516,7 +1484,7 @@ export class Resources {
    * @private
    */
   cleanupTasks_(resource, opt_removePending) {
-    if (resource.getState() == ResourceState_.NOT_LAID_OUT) {
+    if (resource.getState() == ResourceState.NOT_LAID_OUT) {
       // If the layout promise for this resource has not resolved yet, remove
       // it from the task queues to make sure this resource can be rescheduled
       // for layout again later on.
@@ -1532,762 +1500,11 @@ export class Resources {
           request => request.resource != resource);
     }
 
-    if (resource.getState() == ResourceState_.NOT_BUILT && opt_removePending &&
+    if (resource.getState() == ResourceState.NOT_BUILT && opt_removePending &&
         this.pendingBuildResources_) {
       const pendingIndex = this.pendingBuildResources_.indexOf(resource);
       if (pendingIndex != -1) {
         this.pendingBuildResources_.splice(pendingIndex, 1);
-      }
-    }
-  }
-}
-
-
-/**
- * A Resource binding for an AmpElement.
- *
- * Visible for testing only!
- */
-export class Resource {
-
-  /**
-   * @param {number} id
-   * @param {!AmpElement} element
-   * @param {!Resources} resources
-   */
-  constructor(id, element, resources) {
-    /** @private {number} */
-    this.id_ = id;
-
-    /** @export @const {!AmpElement} */
-    this.element = element;
-
-    /** @export @const {string} */
-    this.debugid = element.tagName.toLowerCase() + '#' + id;
-
-    /** @private {!Resources} */
-    this.resources_ = resources;
-
-    /** @private {boolean} */
-    this.blacklisted_ = false;
-
-    /** @const {!AmpElement|undefined|null} */
-    this.owner_ = undefined;
-
-    /** @private {!ResourceState_} */
-    this.state_ = element.isBuilt() ? ResourceState_.NOT_LAID_OUT :
-        ResourceState_.NOT_BUILT;
-
-    /** @private {number} */
-    this.layoutCount_ = 0;
-
-    /** @private {boolean} */
-    this.isFixed_ = false;
-
-    /** @private {!../layout-rect.LayoutRectDef} */
-    this.layoutBox_ = layoutRectLtwh(-10000, -10000, 0, 0);
-
-    /** @private {!../layout-rect.LayoutRectDef} */
-    this.initialLayoutBox_ = this.layoutBox_;
-
-    /** @private {boolean} */
-    this.isMeasureRequested_ = false;
-
-    /** @private {boolean} */
-    this.isInViewport_ = false;
-
-    /** @private {?Promise<undefined>} */
-    this.layoutPromise_ = null;
-
-    /**
-     * Only used in the "runtime off" case when the monitoring code needs to
-     * known when the element is upgraded.
-     * @private {!Function|undefined}
-     */
-    this.onUpgraded_ = undefined;
-
-   /**
-    * Pending change size that was requested but could not be satisfied.
-    * @private {!SizeDef|undefined}
-    */
-    this.pendingChangeSize_ = undefined;
-
-    /** @private @const {!Promise} */
-    this.loadPromise_ = new Promise(resolve => {
-      /** @const  */
-      this.loadPromiseResolve_ = resolve;
-    });
-
-    /** @private {boolean} */
-    this.paused_ = false;
-  }
-
-  /**
-   * Returns resource's ID.
-   * @return {number}
-   */
-  getId() {
-    return this.id_;
-  }
-
-  /**
-   * Returns an owner element or null.
-   * @return {?AmpElement}
-   */
-  getOwner() {
-    if (this.owner_ === undefined) {
-      for (let n = this.element; n; n = n.parentElement) {
-        if (n[OWNER_PROP_]) {
-          this.owner_ = n[OWNER_PROP_];
-          break;
-        }
-      }
-      if (this.owner_ === undefined) {
-        this.owner_ = null;
-      }
-    }
-    return this.owner_;
-  }
-
-  /**
-   * Whether the resource has an owner.
-   * @return {boolean}
-   */
-  hasOwner() {
-    return !!this.getOwner();
-  }
-
-  /**
-   * Returns the resource's element priority.
-   * @return {number}
-   */
-  getPriority() {
-    return this.element.getPriority();
-  }
-
-  /**
-   * Returns the resource's state. See {@link ResourceState_} for details.
-   * @return {!ResourceState_}
-   */
-  getState() {
-    return this.state_;
-  }
-
-  /**
-   * Requests the resource's element to be built. See {@link AmpElement.build}
-   * for details.
-   */
-  build() {
-    if (this.blacklisted_ || !this.element.isUpgraded()) {
-      return;
-    }
-    try {
-      this.element.build();
-    } catch (e) {
-      dev.error(TAG_, 'failed to build:', this.debugid, e);
-      this.blacklisted_ = true;
-      return;
-    }
-
-    if (this.hasBeenMeasured()) {
-      this.state_ = ResourceState_.READY_FOR_LAYOUT;
-    } else {
-      this.state_ = ResourceState_.NOT_LAID_OUT;
-    }
-  }
-
-  /**
-   * Optionally hides or shows the element depending on the media query.
-   */
-  applySizesAndMediaQuery() {
-    this.element.applySizesAndMediaQuery();
-  }
-
-  /**
-   * Instructs the element to change its size and transitions to the state
-   * awaiting the measure and possibly layout.
-   * @param {number|undefined} newHeight
-   * @param {number|undefined} newWidth
-   * @param {function()=} opt_callback A callback function.
-   */
-  changeSize(newHeight, newWidth, opt_callback) {
-    this.element./*OK*/changeSize(newHeight, newWidth);
-    // Schedule for re-layout.
-    if (this.state_ != ResourceState_.NOT_BUILT) {
-      this.state_ = ResourceState_.NOT_LAID_OUT;
-    }
-    if (opt_callback) {
-      opt_callback();
-    }
-  }
-
-  /**
-   * Informs the element that it's either overflown or not.
-   * @param {boolean} overflown
-   * @param {number|undefined} requestedHeight
-   * @param {number|undefined} requestedWidth
-   */
-  overflowCallback(overflown, requestedHeight, requestedWidth) {
-    if (overflown) {
-      this.pendingChangeSize_ = {
-        height: requestedHeight,
-        width: requestedWidth,
-      };
-    }
-    this.element.overflowCallback(overflown, requestedHeight, requestedWidth);
-  }
-
-  /** @private */
-  resetPendingChangeSize() {
-    this.pendingChangeSize_ = undefined;
-  }
-
-  /**
-   * @return {!SizeDef|undefined}
-   */
-  getPendingChangeSize() {
-    return this.pendingChangeSize_;
-  }
-
-  /**
-   * Measures the resource's boundaries. Only allowed for upgraded elements.
-   */
-  measure() {
-    this.isMeasureRequested_ = false;
-    const box = this.resources_.viewport_.getLayoutRect(this.element);
-    // Note that "left" doesn't affect readiness for the layout.
-    if (this.state_ == ResourceState_.NOT_LAID_OUT ||
-          this.layoutBox_.top != box.top ||
-          this.layoutBox_.width != box.width ||
-          this.layoutBox_.height != box.height) {
-
-      if (this.element.isUpgraded() &&
-              this.state_ != ResourceState_.NOT_BUILT &&
-              (this.state_ == ResourceState_.NOT_LAID_OUT ||
-                  this.element.isRelayoutNeeded())) {
-        this.state_ = ResourceState_.READY_FOR_LAYOUT;
-      }
-    }
-    if (!this.hasBeenMeasured()) {
-      this.initialLayoutBox_ = box;
-    }
-    this.layoutBox_ = box;
-
-    // Calculate whether the element is currently is or in `position:fixed`.
-    let isFixed = false;
-    if (this.isDisplayed()) {
-      const win = this.resources_.win;
-      const body = win.document.body;
-      const viewport = this.resources_.viewport_;
-      for (let n = this.element; n && n != body; n = n./*OK*/offsetParent) {
-        if (n.isAlwaysFixed && n.isAlwaysFixed()) {
-          isFixed = true;
-          break;
-        }
-        if (viewport.isDeclaredFixed(n) &&
-                win./*OK*/getComputedStyle(n).position == 'fixed') {
-          isFixed = true;
-          break;
-        }
-      }
-    }
-    this.isFixed_ = isFixed;
-
-    this.element.updateLayoutBox(box);
-  }
-
-  /**
-   * @return {boolean}
-   */
-  isMeasureRequested() {
-    return this.isMeasureRequested_;
-  }
-
-  /**
-   * Checks if the current resource has been measured.
-   * @return {boolean}
-   */
-  hasBeenMeasured() {
-    return this.layoutBox_.top != -10000;
-  }
-
-  /**
-   * Requests the element to be remeasured on the next pass.
-   */
-  requestMeasure() {
-    if (this.state_ == ResourceState_.NOT_BUILT) {
-      // Can't measure unbuilt element.
-      return;
-    }
-    this.isMeasureRequested_ = true;
-  }
-
-  /**
-   * Returns a previously measured layout box.
-   * @return {!../layout-rect.LayoutRectDef}
-   */
-  getLayoutBox() {
-    return this.layoutBox_;
-  }
-
-  /**
-   * Returns the first measured layout box.
-   * @return {!../layout-rect.LayoutRectDef}
-   */
-  getInitialLayoutBox() {
-    return this.initialLayoutBox_;
-  }
-
-  /**
-   * Whether the resource is displayed, i.e. if it has non-zero width and
-   * height.
-   * @return {boolean}
-   */
-  isDisplayed() {
-    return this.layoutBox_.height > 0 && this.layoutBox_.width > 0;
-  }
-
-  /**
-   * Whether the element is fixed according to the latest measurement.
-   * @return {boolean}
-   */
-  isFixed() {
-    return this.isFixed_;
-  }
-
-  /**
-   * Whether the element's layout box overlaps with the specified rect.
-   * @param {!../layout-rect.LayoutRectDef} rect
-   * @return {boolean}
-   */
-  overlaps(rect) {
-    return layoutRectsOverlap(this.layoutBox_, rect);
-  }
-
-  /**
-   * Whether this element can be pre-rendered.
-   * @return {boolean}
-   */
-  prerenderAllowed() {
-    return this.element.prerenderAllowed();
-  }
-
-  /**
-   * Whether this is allowed to render when not in viewport.
-   * @return {boolean}
-   */
-  renderOutsideViewport() {
-    const renders = this.element.renderOutsideViewport();
-    // Boolean interface, element is either always allowed or never allowed to
-    // render outside viewport.
-    if (renders === true || renders === false) {
-      return renders;
-    }
-    // Numeric interface, element is allowed to render outside viewport when it
-    // is within X times the viewport height of the current viewport.
-    const viewportBox = this.resources_.getViewport().getRect();
-    const layoutBox = this.layoutBox_;
-    const scrollDirection = this.resources_.getScrollDirection();
-    const multipler = Math.max(renders, 0);
-    let scrollPenalty = 1;
-    let distance;
-    if (viewportBox.bottom < layoutBox.top) {
-      // Element is below viewport
-      distance = layoutBox.top - viewportBox.bottom;
-
-      // If we're scrolling away from the element
-      if (scrollDirection == -1) {
-        scrollPenalty = 2;
-      }
-    } else if (viewportBox.top > layoutBox.bottom) {
-      // Element is above viewport
-      distance = viewportBox.top - layoutBox.bottom;
-
-      // If we're scrolling away from the element
-      if (scrollDirection == 1) {
-        scrollPenalty = 2;
-      }
-    } else {
-      // Element is in viewport
-      return true;
-    }
-    return distance < viewportBox.height * multipler / scrollPenalty;
-  }
-
-  /**
-   * Sets the resource's state to LAYOUT_SCHEDULED.
-   */
-  layoutScheduled() {
-    this.state_ = ResourceState_.LAYOUT_SCHEDULED;
-  }
-
-  /**
-   * Starts the layout of the resource. Returns the promise that will yield
-   * once layout is complete. Only allowed to be called on a upgraded, built
-   * and displayed element.
-   * @param {boolean} isDocumentVisible
-   * @return {!Promise}
-   */
-  startLayout(isDocumentVisible) {
-    if (this.layoutPromise_) {
-      return this.layoutPromise_;
-    }
-    if (this.state_ == ResourceState_.LAYOUT_COMPLETE) {
-      return Promise.resolve();
-    }
-    if (this.state_ == ResourceState_.LAYOUT_FAILED) {
-      return Promise.reject('already failed');
-    }
-
-    dev.assert(this.state_ != ResourceState_.NOT_BUILT,
-        'Not ready to start layout: %s (%s)', this.debugid, this.state_);
-
-    if (!isDocumentVisible && !this.prerenderAllowed()) {
-      dev.fine(TAG_, 'layout canceled due to non pre-renderable element:',
-          this.debugid, this.state_);
-      this.state_ = ResourceState_.READY_FOR_LAYOUT;
-      return Promise.resolve();
-    }
-
-    if (!this.isInViewport() && !this.renderOutsideViewport()) {
-      dev.fine(TAG_, 'layout canceled due to element not being in viewport:',
-          this.debugid, this.state_);
-      this.state_ = ResourceState_.READY_FOR_LAYOUT;
-      return Promise.resolve();
-    }
-
-    // Double check that the element has not disappeared since scheduling
-    this.measure();
-    if (!this.isDisplayed()) {
-      dev.fine(TAG_, 'layout canceled due to element loosing display:',
-          this.debugid, this.state_);
-      return Promise.resolve();
-    }
-
-    // Not-wanted re-layouts are ignored.
-    if (this.layoutCount_ > 0 && !this.element.isRelayoutNeeded()) {
-      dev.fine(TAG_, 'layout canceled since it wasn\'t requested:',
-          this.debugid, this.state_);
-      this.state_ = ResourceState_.LAYOUT_COMPLETE;
-      return Promise.resolve();
-    }
-
-    dev.fine(TAG_, 'start layout:', this.debugid, 'count:', this.layoutCount_);
-    this.layoutCount_++;
-    this.state_ = ResourceState_.LAYOUT_SCHEDULED;
-
-    this.resources_.framerate_.collect(this.element);
-    let promise;
-    try {
-      promise = this.element.layoutCallback();
-    } catch (e) {
-      return Promise.reject(e);
-    }
-
-    this.layoutPromise_ = promise.then(() => this.layoutComplete_(true),
-        reason => this.layoutComplete_(false, reason));
-    return this.layoutPromise_;
-  }
-
-  /**
-   * @param {boolean} success
-   * @param {*=} opt_reason
-   * @return {!Promise|undefined}
-   */
-  layoutComplete_(success, opt_reason) {
-    this.loadPromiseResolve_();
-    this.layoutPromise_ = null;
-    this.state_ = success ? ResourceState_.LAYOUT_COMPLETE :
-        ResourceState_.LAYOUT_FAILED;
-    if (success) {
-      dev.fine(TAG_, 'layout complete:', this.debugid);
-    } else {
-      dev.fine(TAG_, 'loading failed:', this.debugid, opt_reason);
-      return Promise.reject(opt_reason);
-    }
-  }
-
-  /**
-   * Returns true if the resource layout has not completed or failed.
-   * @return {boolean}
-   * */
-  isLayoutPending() {
-    return this.state_ != ResourceState_.LAYOUT_COMPLETE &&
-        this.state_ != ResourceState_.LAYOUT_FAILED;
-  }
-
-  /**
-   * Returns a promise that is resolved when this resource is laid out
-   * for the first time and the resource was loaded.
-   * @return {!Promise}
-   */
-  loaded() {
-    return this.loadPromise_;
-  }
-
-  /**
-   * Whether the resource is currently visible in the viewport.
-   * @return {boolean}
-   */
-  isInViewport() {
-    return this.isInViewport_;
-  }
-
-  /**
-   * Updates the inViewport state of the element.
-   * @param {boolean} inViewport
-   */
-  setInViewport(inViewport) {
-    if (inViewport == this.isInViewport_) {
-      return;
-    }
-    dev.fine(TAG_, 'inViewport:', this.debugid, inViewport);
-    this.isInViewport_ = inViewport;
-    this.element.viewportCallback(inViewport);
-  }
-
-  /**
-   * Calls element's unlayoutCallback callback and resets state for
-   * relayout in case document becomes active again.
-   */
-  unlayout() {
-    if (this.state_ == ResourceState_.NOT_BUILT ||
-        this.state_ == ResourceState_.NOT_LAID_OUT) {
-      return;
-    }
-    this.setInViewport(false);
-    if (this.element.unlayoutCallback()) {
-      this.element.togglePlaceholder(true);
-      this.state_ = ResourceState_.NOT_LAID_OUT;
-      this.layoutCount_ = 0;
-      this.layoutPromise_ = null;
-    }
-  }
-
-  /**
-   * Returns the task ID for this resource.
-   * @param localId
-   * @returns {string}
-   */
-  getTaskId(localId) {
-    return this.debugid + '#' + localId;
-  }
-
-  /**
-   * Calls element's pauseCallback callback.
-   */
-  pause() {
-    if (this.state_ == ResourceState_.NOT_BUILT || this.paused_) {
-      return;
-    }
-    this.paused_ = true;
-    this.setInViewport(false);
-    this.element.pauseCallback();
-    if (this.element.unlayoutOnPause()) {
-      this.unlayout();
-    }
-  }
-
-  /**
-   * Calls element's resumeCallback callback.
-   */
-  resume() {
-    if (this.state_ == ResourceState_.NOT_BUILT || !this.paused_) {
-      return;
-    }
-    this.paused_ = false;
-    this.element.resumeCallback();
-  }
-
-  /**
-   * Called when a previously visible element is no longer displayed.
-   */
-  unload() {
-    this.pause();
-    this.unlayout();
-  }
-
-  /**
-   * Only allowed in dev mode when runtime is turned off. Performs all steps
-   * necessary to render an element.
-   * @return {!Promise}
-   * @export
-   */
-  forceAll() {
-    dev.assert(!this.resources_.isRuntimeOn_);
-    let p = Promise.resolve();
-    if (this.state_ == ResourceState_.NOT_BUILT) {
-      if (!this.element.isUpgraded()) {
-        p = p.then(() => {
-          return new Promise(resolve => {
-            this.onUpgraded_ = resolve;
-          });
-        });
-      }
-      p = p.then(() => {
-        this.onUpgraded_ = undefined;
-        this.build(true);
-      });
-    }
-    return p.then(() => {
-      this.applySizesAndMediaQuery();
-      this.measure();
-      if (this.layoutPromise_) {
-        return this.layoutPromise_;
-      }
-      if (this.state_ == ResourceState_.LAYOUT_COMPLETE ||
-              this.state_ == ResourceState_.LAYOUT_FAILED ||
-              this.layoutCount_ > 0) {
-        return;
-      }
-      if (!this.isDisplayed()) {
-        return;
-      }
-      this.layoutCount_++;
-      return this.element.layoutCallback();
-    });
-  }
-}
-
-
-/**
- * A scheduling queue for Resources.
- *
- * Visible only for testing!
- *
- * @private
- */
-export class TaskQueue_ {
-
-  constructor() {
-    /** @private @const {!Array<!TaskDef>} */
-    this.tasks_ = [];
-
-    /** @private @const {!Object<string, !TaskDef>} */
-    this.taskIdMap_ = {};
-
-    /** @private {!time} */
-    this.lastEnqueueTime_ = 0;
-
-    /** @private {!time} */
-    this.lastDequeueTime_ = 0;
-  }
-
-  /**
-   * Size of the queue.
-   * @return {number}
-   */
-  getSize() {
-    return this.tasks_.length;
-  }
-
-  /**
-   * Last time a task was enqueued.
-   * @return {!time}
-   */
-  getLastEnqueueTime() {
-    return this.lastEnqueueTime_;
-  }
-
-  /**
-   * Last time a task was dequeued.
-   * @return {!time}
-   */
-  getLastDequeueTime() {
-    return this.lastDequeueTime_;
-  }
-
-  /**
-   * Returns the task with the specified ID or null.
-   * @param {string} taskId
-   * @return {?TaskDef}
-   */
-  getTaskById(taskId) {
-    return this.taskIdMap_[taskId] || null;
-  }
-
-  /**
-   * Enqueues the task. If the task is already in the queue, the error is
-   * thrown.
-   * @param {!TaskDef} task
-   */
-  enqueue(task) {
-    dev.assert(!this.taskIdMap_[task.id], 'Task already enqueued: %s', task.id);
-    this.tasks_.push(task);
-    this.taskIdMap_[task.id] = task;
-    this.lastEnqueueTime_ = timer.now();
-  }
-
-  /**
-   * Dequeues the task and returns "true" if dequeueing is successful. Otherwise
-   * returns "false", e.g. when this task is not currently enqueued.
-   * @param {!TaskDef} task
-   * @return {boolean}
-   */
-  dequeue(task) {
-    const existing = this.taskIdMap_[task.id];
-    const dequeued = this.removeAtIndex(task, this.tasks_.indexOf(existing));
-    if (!dequeued) {
-      return false;
-    }
-    this.lastDequeueTime_ = timer.now();
-    return true;
-  }
-
-  /**
-   * Returns the task with the minimal score based on the provided scoring
-   * callback.
-   * @param {function(!TaskDef):number} scorer
-   * @return {?TaskDef}
-   */
-  peek(scorer) {
-    let minScore = 1e6;
-    let minTask = null;
-    for (let i = 0; i < this.tasks_.length; i++) {
-      const task = this.tasks_[i];
-      const score = scorer(task);
-      if (score < minScore) {
-        minScore = score;
-        minTask = task;
-      }
-    }
-    return minTask;
-  }
-
-  /**
-   * Iterates over all tasks in queue in the insertion order.
-   * @param {function(!TaskDef)} callback
-   */
-  forEach(callback) {
-    this.tasks_.forEach(callback);
-  }
-
-  /**
-   * Removes the task and returns "true" if dequeueing is successful. Otherwise
-   * returns "false", e.g. when this task is not currently enqueued.
-   * @param {!TaskDef} task
-   * @param {number} index of the task to remove.
-   * @return {boolean}
-   */
-  removeAtIndex(task, index) {
-    const existing = this.taskIdMap_[task.id];
-    if (!existing || this.tasks_[index] != existing) {
-      return false;
-    }
-    this.tasks_.splice(index, 1);
-    delete this.taskIdMap_[task.id];
-    return true;
-  }
-
-  /**
-   * Removes tasks in queue that pass the callback test.
-   * @param {function(!TaskDef):boolean} callback Return true to remove the task.
-   */
-  purge(callback) {
-    let index = this.tasks_.length;
-    while (index--) {
-      if (callback(this.tasks_[index])) {
-        this.removeAtIndex(this.tasks_[index], index);
       }
     }
   }
@@ -2303,63 +1520,6 @@ function elements_(elements) {
 }
 
 
-/**
- * Resource state.
- *
- * Visible for testing only!
- *
- * @enum {number}
- * @private
- */
-export const ResourceState_ = {
-  /**
-   * The resource has not been built yet. Measures, layouts, preloads or
-   * viewport signals are not allowed.
-   */
-  NOT_BUILT: 0,
-
-  /**
-   * The resource has been built, but not measured yet and not yet ready
-   * for layout.
-   */
-  NOT_LAID_OUT: 1,
-
-  /**
-   * The resource has been built and measured and ready for layout.
-   */
-  READY_FOR_LAYOUT: 2,
-
-  /**
-   * The resource is currently scheduled for layout.
-   */
-  LAYOUT_SCHEDULED: 3,
-
-  /**
-   * The resource has been laid out.
-   */
-  LAYOUT_COMPLETE: 4,
-
-  /**
-   * The latest resource's layout failed.
-   */
-  LAYOUT_FAILED: 5,
-};
-
-
-/**
- * The internal structure for the task.
- * @typedef {{
- *   id: string,
- *   resource: !Resource,
- *   priority: number,
- *   callback: function(boolean),
- *   scheduleTime: time,
- *   startTime: time,
- *   promise: (!Promise|undefined)
- * }}
- * @private
- */
-let TaskDef;
 
 /**
  * The internal structure of a ChangeHeightRequest.
