+++ test/functional/test-resources.js
@@ -14,14 +14,9 @@
  * limitations under the License.
  */
 
-import {
-  Resource,
-  ResourceState_,
-  Resources,
-  TaskQueue_,
-} from '../../src/service/resources-impl';
+import {Resources} from '../../src/service/resources-impl';
+import {Resource, ResourceState} from '../../src/service/resource';
 import {VisibilityState} from '../../src/visibility-state';
-import {dev} from '../../src/log';
 import {layoutRectLtwh} from '../../src/layout-rect';
 import * as sinon from 'sinon';
 
@@ -204,7 +199,7 @@ describe('Resources', () => {
   it('should not schedule non-prerenderable resource when' +
         ' document is hidden', () => {
     const resource = {
-      getState: () => ResourceState_.READY_FOR_LAYOUT,
+      getState: () => ResourceState.READY_FOR_LAYOUT,
       isDisplayed: () => true,
       isFixed: () => false,
       isInViewport: () => true,
@@ -223,7 +218,7 @@ describe('Resources', () => {
   it('should schedule prerenderable resource when' +
         ' document is hidden', () => {
     const resource = {
-      getState: () => ResourceState_.READY_FOR_LAYOUT,
+      getState: () => ResourceState.READY_FOR_LAYOUT,
       isDisplayed: () => true,
       isFixed: () => false,
       isInViewport: () => true,
@@ -245,7 +240,7 @@ describe('Resources', () => {
   it('should not schedule non-renderOutsideViewport resource when' +
         ' resource is not visible', () => {
     const resource = {
-      getState: () => ResourceState_.READY_FOR_LAYOUT,
+      getState: () => ResourceState.READY_FOR_LAYOUT,
       isDisplayed: () => true,
       isFixed: () => false,
       isInViewport: () => false,
@@ -260,7 +255,7 @@ describe('Resources', () => {
   it('should schedule renderOutsideViewport resource when' +
         ' resource is not visible', () => {
     const resource = {
-      getState: () => ResourceState_.READY_FOR_LAYOUT,
+      getState: () => ResourceState.READY_FOR_LAYOUT,
       isDisplayed: () => true,
       isFixed: () => false,
       isInViewport: () => false,
@@ -325,7 +320,7 @@ describe('Resources schedulePause', () => {
   function createElementWithResource(id) {
     const element = createElement();
     const resource = new Resource(id, element, resources);
-    resource.state_ = ResourceState_.LAYOUT_COMPLETE;
+    resource.state_ = ResourceState.LAYOUT_COMPLETE;
     resource.element['__AMP__RESOURCE'] = resource;
     return [element, resource];
   }
@@ -441,7 +436,7 @@ describe('Resources schedulePreload', () => {
   function createElementWithResource(id) {
     const element = createElement();
     const resource = new Resource(id, element, resources);
-    resource.state_ = ResourceState_.READY_FOR_LAYOUT;
+    resource.state_ = ResourceState.READY_FOR_LAYOUT;
     resource.element['__AMP__RESOURCE'] = resource;
     resource.measure = sandbox.spy();
     resource.isDisplayed = () => true;
@@ -551,7 +546,7 @@ describe('Resources discoverWork', () => {
 
   function createResource(id, rect) {
     const resource = new Resource(id, createElement(rect), resources);
-    resource.state_ = ResourceState_.READY_FOR_LAYOUT;
+    resource.state_ = ResourceState.READY_FOR_LAYOUT;
     resource.layoutBox_ = rect;
     return resource;
   }
@@ -595,8 +590,8 @@ describe('Resources discoverWork', () => {
   });
 
   it('should NOT rerender anything', () => {
-    resource1.state_ = ResourceState_.LAYOUT_COMPLETE;
-    resource2.state_ = ResourceState_.LAYOUT_COMPLETE;
+    resource1.state_ = ResourceState.LAYOUT_COMPLETE;
+    resource2.state_ = ResourceState.LAYOUT_COMPLETE;
     resources.visible_ = true;
     sandbox.stub(resources.viewer_, 'getVisibilityState').returns(
       VisibilityState.VISIBLE
@@ -610,8 +605,8 @@ describe('Resources discoverWork', () => {
   });
 
   it('should re-render from requested position', () => {
-    resource1.state_ = ResourceState_.LAYOUT_COMPLETE;
-    resource2.state_ = ResourceState_.LAYOUT_COMPLETE;
+    resource1.state_ = ResourceState.LAYOUT_COMPLETE;
+    resource2.state_ = ResourceState.LAYOUT_COMPLETE;
     resource1.element.getBoundingClientRect =
         () => layoutRectLtwh(10, 10, 100, 101);
     resource2.element.getBoundingClientRect =
@@ -630,8 +625,8 @@ describe('Resources discoverWork', () => {
     expect(resources.relayoutTop_).to.equal(-1);
     expect(resources.queue_.getSize()).to.equal(1);
     expect(resources.queue_.tasks_[0].resource).to.equal(resource2);
-    expect(resource1.state_).to.equal(ResourceState_.LAYOUT_COMPLETE);
-    expect(resource2.state_).to.equal(ResourceState_.LAYOUT_SCHEDULED);
+    expect(resource1.state_).to.equal(ResourceState.LAYOUT_COMPLETE);
+    expect(resource2.state_).to.equal(ResourceState.LAYOUT_SCHEDULED);
   });
 
   it('should prerender only one screen with prerenderSize = 1', () => {
@@ -664,8 +659,8 @@ describe('Resources discoverWork', () => {
   });
 
   it('should remeasure when requested and scheduled unloads', () => {
-    resource1.state_ = ResourceState_.LAYOUT_COMPLETE;
-    resource2.state_ = ResourceState_.LAYOUT_COMPLETE;
+    resource1.state_ = ResourceState.LAYOUT_COMPLETE;
+    resource2.state_ = ResourceState.LAYOUT_COMPLETE;
     resources.visible_ = true;
     sandbox.stub(resources.viewer_, 'getVisibilityState').returns(
       VisibilityState.VISIBLE
@@ -710,7 +705,7 @@ describe('Resources discoverWork', () => {
 
   it('should eject stale tasks when element unloaded', () => {
     const pendingResource = createResource(5, layoutRectLtwh(0, 0, 0, 0));
-    pendingResource.state_ = ResourceState_.NOT_BUILT;
+    pendingResource.state_ = ResourceState.NOT_BUILT;
     resources.pendingBuildResources_ = [pendingResource];
     resources.visible_ = true;
     // Don't resolve layout - immulating DOM being removed and load
@@ -802,7 +797,7 @@ describe('Resources changeSize', () => {
   function createResource(id, rect) {
     const resource = new Resource(id, createElement(rect), resources);
     resource.element['__AMP__RESOURCE'] = resource;
-    resource.state_ = ResourceState_.READY_FOR_LAYOUT;
+    resource.state_ = ResourceState.READY_FOR_LAYOUT;
     resource.layoutBox_ = rect;
     resource.changeSize = sandbox.spy();
     return resource;
@@ -1190,7 +1185,7 @@ describe('Resources mutateElement', () => {
         createElement(rect, /* isAmp */ true),
         resources);
     resource.element['__AMP__RESOURCE'] = resource;
-    resource.state_ = ResourceState_.READY_FOR_LAYOUT;
+    resource.state_ = ResourceState.READY_FOR_LAYOUT;
     resource.layoutBox_ = rect;
     resource.changeSize = sandbox.spy();
     return resource;
@@ -1320,1291 +1315,6 @@ describe('Resources mutateElement', () => {
 });
 
 
-describe('Resources.TaskQueue', () => {
-
-  let sandbox;
-  let clock;
-  let queue;
-
-  beforeEach(() => {
-    sandbox = sinon.sandbox.create();
-    clock = sandbox.useFakeTimers();
-    queue = new TaskQueue_();
-  });
-
-  afterEach(() => {
-    sandbox.restore();
-  });
-
-  it('should enqueue and dequeue', () => {
-    clock.tick(1000);
-    expect(queue.getSize()).to.equal(0);
-    expect(queue.getLastEnqueueTime()).to.equal(0);
-    expect(queue.getLastDequeueTime()).to.equal(0);
-
-    queue.enqueue({id: '1'});
-    expect(queue.getTaskById('1').id).to.equal('1');
-    expect(queue.getSize()).to.equal(1);
-    expect(queue.getLastEnqueueTime()).to.equal(1000);
-    expect(queue.getLastDequeueTime()).to.equal(0);
-
-    expect(() => {
-      queue.enqueue({id: '1'});
-    }).to.throw(/Task already enqueued/);
-
-    queue.dequeue({id: '1'});
-    expect(queue.getTaskById('1')).to.equal(null);
-    expect(queue.getSize()).to.equal(0);
-    expect(queue.getLastEnqueueTime()).to.equal(1000);
-    expect(queue.getLastDequeueTime()).to.equal(1000);
-  });
-
-  it('should perform score-based peek', () => {
-    queue.enqueue({id: 'A', v: 0});
-    queue.enqueue({id: 'B', v: 2});
-    queue.enqueue({id: 'C', v: 1});
-
-    const task = queue.peek(task => 10 - task.v);
-    expect(task.id).to.equal('B');
-  });
-});
-
-
-describe('Resources.Resource', () => {
-
-  let sandbox;
-  let element;
-  let elementMock;
-  let resources;
-  let resource;
-  let viewportMock;
-
-  beforeEach(() => {
-    sandbox = sinon.sandbox.create();
-
-    element = {
-      tagName: 'AMP-AD',
-      isBuilt: () => false,
-      isUpgraded: () => false,
-      prerenderAllowed: () => false,
-      renderOutsideViewport: () => true,
-      build: () => false,
-      getBoundingClientRect: () => null,
-      updateLayoutBox: () => {},
-      isRelayoutNeeded: () => false,
-      layoutCallback: () => {},
-      changeSize: () => {},
-      unlayoutOnPause: () => false,
-      unlayoutCallback: () => true,
-      pauseCallback: () => false,
-      resumeCallback: () => false,
-      viewportCallback: () => {},
-      togglePlaceholder: () => sandbox.spy(),
-      getPriority: () => 2,
-    };
-    elementMock = sandbox.mock(element);
-
-    resources = new Resources(window);
-    resource = new Resource(1, element, resources);
-    viewportMock = sandbox.mock(resources.viewport_);
-  });
-
-  afterEach(() => {
-    viewportMock.verify();
-    elementMock.verify();
-    sandbox.restore();
-  });
-
-  it('should initialize correctly', () => {
-    expect(resource.getId()).to.equal(1);
-    expect(resource.debugid).to.equal('amp-ad#1');
-    expect(resource.getPriority()).to.equal(2);
-    expect(resource.getState()).to.equal(ResourceState_.NOT_BUILT);
-    expect(resource.getLayoutBox().width).to.equal(0);
-    expect(resource.getLayoutBox().height).to.equal(0);
-    expect(resource.isInViewport()).to.equal(false);
-  });
-
-  it('should initialize correctly when already built', () => {
-    elementMock.expects('isBuilt').returns(true).once();
-    expect(new Resource(1, element).getState()).to.equal(
-        ResourceState_.NOT_LAID_OUT);
-  });
-
-  it('should not build before upgraded', () => {
-    elementMock.expects('isUpgraded').returns(false).atLeast(1);
-    elementMock.expects('build').never();
-
-    resource.build();
-    expect(resource.getState()).to.equal(ResourceState_.NOT_BUILT);
-  });
-
-
-  it('should build after upgraded', () => {
-    elementMock.expects('isUpgraded').returns(true).atLeast(1);
-    elementMock.expects('build').once();
-    resource.build();
-    expect(resource.getState()).to.equal(ResourceState_.NOT_LAID_OUT);
-  });
-
-  it('should blacklist on build failure', () => {
-    elementMock.expects('isUpgraded').returns(true).atLeast(1);
-    elementMock.expects('build').throws('Failed').once();
-    resource.build();
-    expect(resource.blacklisted_).to.equal(true);
-    expect(resource.getState()).to.equal(ResourceState_.NOT_BUILT);
-  });
-
-  it('should mark as ready for layout if already measured', () => {
-    elementMock.expects('isUpgraded').returns(true).atLeast(1);
-    elementMock.expects('build').once();
-    const stub = sandbox.stub(resource, 'hasBeenMeasured').returns(true);
-    resource.build(false);
-    expect(stub.calledOnce).to.be.true;
-    expect(resource.getState()).to.equal(ResourceState_.READY_FOR_LAYOUT);
-  });
-
-  it('should mark as not laid out if not yet measured', () => {
-    elementMock.expects('isUpgraded').returns(true).atLeast(1);
-    elementMock.expects('build').once();
-    const stub = sandbox.stub(resource, 'hasBeenMeasured').returns(false);
-    resource.build(false);
-    expect(stub.calledOnce).to.be.true;
-    expect(resource.getState()).to.equal(ResourceState_.NOT_LAID_OUT);
-  });
-
-  it('should allow to measure when not upgraded', () => {
-    elementMock.expects('isUpgraded').returns(false).atLeast(1);
-    resource.resources_ = {
-      win: window,
-      viewport_: {
-        getLayoutRect() {
-          return layoutRectLtwh(0, 100, 300, 100);
-        },
-        isDeclaredFixed() {
-          return false;
-        },
-      },
-    };
-    expect(() => {
-      resource.measure();
-    }).to.not.throw();
-    expect(resource.getLayoutBox()).to.eql(layoutRectLtwh(0, 100, 300, 100));
-  });
-
-  it('should allow measure even when not built', () => {
-    elementMock.expects('isUpgraded').returns(true).atLeast(1);
-    elementMock.expects('getBoundingClientRect').returns(
-        layoutRectLtwh(0, 0, 0, 0)).once();
-    resource.measure();
-    expect(resource.getState()).to.equal(ResourceState_.NOT_BUILT);
-    expect(resource.isFixed()).to.be.false;
-  });
-
-  it('should measure and update state', () => {
-    elementMock.expects('isUpgraded').returns(true).atLeast(1);
-    elementMock.expects('build').once();
-    resource.build();
-
-    elementMock.expects('getBoundingClientRect')
-        .returns({left: 11, top: 12, width: 111, height: 222})
-        .once();
-    elementMock.expects('updateLayoutBox')
-        .withExactArgs(sinon.match(data => {
-          return data.width == 111 && data.height == 222;
-        }))
-        .once();
-    resource.measure();
-    expect(resource.getState()).to.equal(ResourceState_.READY_FOR_LAYOUT);
-    expect(resource.getLayoutBox().left).to.equal(11);
-    expect(resource.getLayoutBox().top).to.equal(12);
-    expect(resource.getLayoutBox().width).to.equal(111);
-    expect(resource.getLayoutBox().height).to.equal(222);
-    expect(resource.isFixed()).to.be.false;
-  });
-
-  it('should update initial box only on first measure', () => {
-    elementMock.expects('isUpgraded').returns(true).atLeast(1);
-    elementMock.expects('build').once();
-    resource.build();
-
-    element.getBoundingClientRect = () =>
-        ({left: 11, top: 12, width: 111, height: 222});
-    resource.measure();
-    expect(resource.getLayoutBox().top).to.equal(12);
-    expect(resource.getInitialLayoutBox().top).to.equal(12);
-
-    element.getBoundingClientRect = () =>
-        ({left: 11, top: 22, width: 111, height: 222});
-    resource.measure();
-    expect(resource.getLayoutBox().top).to.equal(22);
-    expect(resource.getInitialLayoutBox().top).to.equal(12);
-  });
-
-  it('should noop request measure when not built', () => {
-    expect(resource.isMeasureRequested()).to.be.false;
-    elementMock.expects('getBoundingClientRect').never();
-    resource.requestMeasure();
-    expect(resource.isMeasureRequested()).to.be.false;
-  });
-
-  it('should request measure when built', () => {
-    expect(resource.isMeasureRequested()).to.be.false;
-    elementMock.expects('getBoundingClientRect').never();
-    resource.state_ = ResourceState_.READY_FOR_LAYOUT;
-    resource.requestMeasure();
-    expect(resource.isMeasureRequested()).to.be.true;
-  });
-
-  it('should always layout if has not been laid out before', () => {
-    elementMock.expects('isUpgraded').returns(true).atLeast(1);
-    resource.state_ = ResourceState_.NOT_LAID_OUT;
-    resource.layoutBox_ = {left: 11, top: 12, width: 111, height: 222};
-
-    elementMock.expects('getBoundingClientRect')
-        .returns(resource.layoutBox_).once();
-    resource.measure();
-    expect(resource.getState()).to.equal(ResourceState_.READY_FOR_LAYOUT);
-  });
-
-  it('should not relayout if has box has not changed', () => {
-    resource.state_ = ResourceState_.LAYOUT_COMPLETE;
-    resource.layoutBox_ = {left: 11, top: 12, width: 111, height: 222};
-
-    // Left is not part of validation.
-    elementMock.expects('getBoundingClientRect')
-        .returns({left: 11 + 10, top: 12, width: 111, height: 222}).once();
-    resource.measure();
-    expect(resource.getState()).to.equal(ResourceState_.LAYOUT_COMPLETE);
-    expect(resource.getLayoutBox().left).to.equal(11 + 10);
-  });
-
-  it('should not relayout if box changed but element didn\'t opt in', () => {
-    elementMock.expects('isUpgraded').returns(true).atLeast(1);
-    resource.state_ = ResourceState_.LAYOUT_COMPLETE;
-    resource.layoutBox_ = {left: 11, top: 12, width: 111, height: 222};
-
-    // Width changed.
-    elementMock.expects('getBoundingClientRect')
-        .returns({left: 11, top: 12, width: 111 + 10, height: 222}).once();
-    elementMock.expects('isRelayoutNeeded').returns(false).atLeast(1);
-    resource.measure();
-    expect(resource.getState()).to.equal(ResourceState_.LAYOUT_COMPLETE);
-    expect(resource.getLayoutBox().width).to.equal(111 + 10);
-  });
-
-  it('should relayout if box changed when element opted in', () => {
-    elementMock.expects('isUpgraded').returns(true).atLeast(1);
-    resource.state_ = ResourceState_.LAYOUT_COMPLETE;
-    resource.layoutBox_ = {left: 11, top: 12, width: 111, height: 222};
-
-    // Width changed.
-    elementMock.expects('getBoundingClientRect')
-        .returns({left: 11, top: 12, width: 111 + 10, height: 222}).once();
-    elementMock.expects('isRelayoutNeeded').returns(true).atLeast(1);
-    resource.measure();
-    expect(resource.getState()).to.equal(ResourceState_.READY_FOR_LAYOUT);
-    expect(resource.getLayoutBox().width).to.equal(111 + 10);
-  });
-
-  it('should calculate NOT fixed for non-displayed elements', () => {
-    elementMock.expects('isUpgraded').returns(true).atLeast(1);
-    elementMock.expects('getBoundingClientRect').returns(
-        layoutRectLtwh(0, 0, 0, 0)).once();
-    element.isAlwaysFixed = () => true;
-    resource.measure();
-    expect(resource.isFixed()).to.be.false;
-  });
-
-  it('should calculate fixed for always-fixed parent', () => {
-    elementMock.expects('isUpgraded').returns(true).atLeast(1);
-    elementMock.expects('getBoundingClientRect').returns(
-        layoutRectLtwh(0, 0, 10, 10)).once();
-    element.offsetParent = {
-      isAlwaysFixed: () => true,
-    };
-    resource.measure();
-    expect(resource.isFixed()).to.be.true;
-  });
-
-  it('should calculate fixed for fixed-style parent', () => {
-    elementMock.expects('isUpgraded').returns(true).atLeast(1);
-    elementMock.expects('getBoundingClientRect').returns(
-        layoutRectLtwh(0, 0, 10, 10)).once();
-    const fixedParent = document.createElement('div');
-    fixedParent.style.position = 'fixed';
-    document.body.appendChild(fixedParent);
-    element.offsetParent = fixedParent;
-    viewportMock.expects('isDeclaredFixed')
-        .withExactArgs(element)
-        .returns(false)
-        .once();
-    viewportMock.expects('isDeclaredFixed')
-        .withExactArgs(fixedParent)
-        .returns(true)
-        .once();
-    resource.measure();
-    expect(resource.isFixed()).to.be.true;
-  });
-
-
-  it('should ignore startLayout if already completed or failed or going',
-        () => {
-          elementMock.expects('layoutCallback').never();
-
-          resource.state_ = ResourceState_.LAYOUT_COMPLETE;
-          resource.startLayout(true);
-
-          resource.state_ = ResourceState_.LAYOUT_FAILED;
-          resource.startLayout(true);
-
-          resource.state_ = ResourceState_.READY_FOR_LAYOUT;
-          resource.layoutPromise_ = {};
-          resource.startLayout(true);
-        });
-
-  it('should fail startLayout if not built', () => {
-    elementMock.expects('layoutCallback').never();
-
-    resource.state_ = ResourceState_.NOT_BUILT;
-    expect(() => {
-      resource.startLayout(true);
-    }).to.throw(/Not ready to start layout/);
-  });
-
-  it('should ignore startLayout if not visible', () => {
-    elementMock.expects('isUpgraded').returns(true).atLeast(1);
-    elementMock.expects('getBoundingClientRect')
-        .returns({left: 1, top: 1, width: 1, height: 1}).once();
-
-    elementMock.expects('layoutCallback').never();
-
-    resource.state_ = ResourceState_.READY_FOR_LAYOUT;
-    resource.layoutBox_ = {left: 11, top: 12, width: 0, height: 0};
-    resource.startLayout(true);
-  });
-
-  it('should force startLayout for first layout', () => {
-    elementMock.expects('isUpgraded').returns(true).atLeast(1);
-    elementMock.expects('getBoundingClientRect')
-        .returns({left: 1, top: 1, width: 1, height: 1}).once();
-
-    elementMock.expects('layoutCallback').returns(Promise.resolve()).once();
-
-    resource.state_ = ResourceState_.READY_FOR_LAYOUT;
-    resource.layoutBox_ = {left: 11, top: 12, width: 10, height: 10};
-    resource.startLayout(true);
-    expect(resource.getState()).to.equal(ResourceState_.LAYOUT_SCHEDULED);
-  });
-
-  it('should ignore startLayout for re-layout when not opt-in', () => {
-    elementMock.expects('isUpgraded').returns(true).atLeast(1);
-    elementMock.expects('getBoundingClientRect')
-        .returns({left: 1, top: 1, width: 1, height: 1}).once();
-
-    elementMock.expects('layoutCallback').never();
-
-    resource.state_ = ResourceState_.READY_FOR_LAYOUT;
-    resource.layoutBox_ = {left: 11, top: 12, width: 10, height: 10};
-    resource.layoutCount_ = 1;
-    elementMock.expects('isRelayoutNeeded').returns(false).atLeast(1);
-    resource.startLayout(true);
-    expect(resource.getState()).to.equal(ResourceState_.LAYOUT_COMPLETE);
-  });
-
-  it('should force startLayout for re-layout when opt-in', () => {
-    elementMock.expects('isUpgraded').returns(true).atLeast(1);
-    elementMock.expects('getBoundingClientRect')
-        .returns({left: 1, top: 1, width: 1, height: 1}).once();
-
-    elementMock.expects('layoutCallback').returns(Promise.resolve()).once();
-
-    resource.state_ = ResourceState_.READY_FOR_LAYOUT;
-    resource.layoutBox_ = {left: 11, top: 12, width: 10, height: 10};
-    resource.layoutCount_ = 1;
-    elementMock.expects('isRelayoutNeeded').returns(true).atLeast(1);
-    resource.startLayout(true);
-    expect(resource.getState()).to.equal(ResourceState_.LAYOUT_SCHEDULED);
-  });
-
-  it('should ignore startLayout when document is hidden' +
-        ' and prerender not allowed', () => {
-    elementMock.expects('isUpgraded').returns(true).atLeast(0);
-    elementMock.expects('getBoundingClientRect')
-        .returns({left: 1, top: 1, width: 1, height: 1}).atLeast(0);
-    elementMock.expects('prerenderAllowed').returns(false).atLeast(1);
-
-    elementMock.expects('layoutCallback').never();
-
-    resource.state_ = ResourceState_.READY_FOR_LAYOUT;
-    resource.layoutBox_ = {left: 11, top: 12, width: 10, height: 10};
-    resource.layoutCount_ = 0;
-    resource.startLayout(false);
-    expect(resource.getState()).to.equal(ResourceState_.READY_FOR_LAYOUT);
-  });
-
-  it('should proceed startLayout when document is hidden' +
-        ' and prerender is allowed', () => {
-    elementMock.expects('isUpgraded').returns(true).atLeast(0);
-    elementMock.expects('getBoundingClientRect')
-        .returns({left: 1, top: 1, width: 1, height: 1}).atLeast(0);
-    elementMock.expects('prerenderAllowed').returns(true).atLeast(1);
-
-    elementMock.expects('layoutCallback').returns(Promise.resolve()).once();
-
-    resource.state_ = ResourceState_.READY_FOR_LAYOUT;
-    resource.layoutBox_ = {left: 11, top: 12, width: 10, height: 10};
-    resource.layoutCount_ = 0;
-    resource.startLayout(false);
-    expect(resource.getState()).to.equal(ResourceState_.LAYOUT_SCHEDULED);
-  });
-
-
-  it('should complete startLayout', () => {
-    elementMock.expects('isUpgraded').returns(true).atLeast(1);
-    elementMock.expects('getBoundingClientRect')
-        .returns({left: 1, top: 1, width: 1, height: 1}).once();
-
-    elementMock.expects('layoutCallback').returns(Promise.resolve()).once();
-
-    resource.state_ = ResourceState_.READY_FOR_LAYOUT;
-    resource.layoutBox_ = {left: 11, top: 12, width: 10, height: 10};
-    const loaded = resource.loaded();
-    const promise = resource.startLayout(true);
-    expect(resource.layoutPromise_).to.not.equal(null);
-    expect(resource.getState()).to.equal(ResourceState_.LAYOUT_SCHEDULED);
-
-    return promise.then(() => {
-      expect(resource.getState()).to.equal(ResourceState_.LAYOUT_COMPLETE);
-      expect(resource.layoutPromise_).to.equal(null);
-      return loaded;  // Just making sure this doesn't time out.
-    });
-  });
-
-  it('should fail startLayout', () => {
-    elementMock.expects('isUpgraded').returns(true).atLeast(1);
-    elementMock.expects('getBoundingClientRect')
-        .returns({left: 1, top: 1, width: 1, height: 1}).once();
-
-    elementMock.expects('layoutCallback').returns(Promise.reject()).once();
-
-    resource.state_ = ResourceState_.READY_FOR_LAYOUT;
-    resource.layoutBox_ = {left: 11, top: 12, width: 10, height: 10};
-    const promise = resource.startLayout(true);
-    expect(resource.layoutPromise_).to.not.equal(null);
-    expect(resource.getState()).to.equal(ResourceState_.LAYOUT_SCHEDULED);
-
-    return promise.then(() => {
-      fail('should not be here');
-    }, () => {
-      expect(resource.getState()).to.equal(ResourceState_.LAYOUT_FAILED);
-      expect(resource.layoutPromise_).to.equal(null);
-    });
-  });
-
-  it('should change size and update state', () => {
-    resource.state_ = ResourceState_.READY_FOR_LAYOUT;
-    elementMock.expects('changeSize').withExactArgs(111, 222).once();
-    resource.changeSize(111, 222);
-    expect(resource.getState()).to.equal(ResourceState_.NOT_LAID_OUT);
-  });
-
-  it('should change size but not state', () => {
-    resource.state_ = ResourceState_.NOT_BUILT;
-    elementMock.expects('changeSize').withExactArgs(111, 222).once();
-    resource.changeSize(111, 222);
-    expect(resource.getState()).to.equal(ResourceState_.NOT_BUILT);
-  });
-
-
-  describe('setInViewport', () => {
-    it('should call viewportCallback when not built', () => {
-      resource.state_ = ResourceState_.NOT_BUILT;
-      elementMock.expects('viewportCallback').withExactArgs(true).once();
-      resource.setInViewport(true);
-      expect(resource.isInViewport()).to.equal(true);
-    });
-
-    it('should call viewportCallback when built', () => {
-      resource.state_ = ResourceState_.LAYOUT_COMPLETE;
-      elementMock.expects('viewportCallback').withExactArgs(true).once();
-      resource.setInViewport(true);
-      expect(resource.isInViewport()).to.equal(true);
-    });
-
-    it('should call viewportCallback only once', () => {
-      resource.state_ = ResourceState_.LAYOUT_COMPLETE;
-      elementMock.expects('viewportCallback').withExactArgs(true).once();
-      resource.setInViewport(true);
-      resource.setInViewport(true);
-      resource.setInViewport(true);
-    });
-  });
-
-
-  describe('unlayoutCallback', () => {
-    it('should NOT call unlayoutCallback on unbuilt element', () => {
-      resource.state_ = ResourceState_.NOT_BUILT;
-      elementMock.expects('viewportCallback').never();
-      elementMock.expects('unlayoutCallback').never();
-      resource.unlayout();
-      expect(resource.getState()).to.equal(ResourceState_.NOT_BUILT);
-    });
-
-    it('should call unlayoutCallback on built element and update state',
-        () => {
-          resource.state_ = ResourceState_.LAYOUT_COMPLETE;
-          elementMock.expects('unlayoutCallback').returns(true).once();
-          elementMock.expects('togglePlaceholder').withArgs(true).once();
-          resource.unlayout();
-          expect(resource.getState()).to.equal(ResourceState_.NOT_LAID_OUT);
-        });
-
-    it('updated state should bypass isRelayoutNeeded', () => {
-      resource.state_ = ResourceState_.LAYOUT_COMPLETE;
-      elementMock.expects('unlayoutCallback').returns(true).once();
-      elementMock.expects('togglePlaceholder').withArgs(true).once();
-      elementMock.expects('isUpgraded').returns(true).atLeast(1);
-      elementMock.expects('getBoundingClientRect')
-          .returns({left: 1, top: 1, width: 1, height: 1}).once();
-
-      resource.unlayout();
-
-      elementMock.expects('layoutCallback').returns(Promise.resolve()).once();
-      resource.startLayout(true);
-    });
-
-    it('should call unlayoutCallback on built element' +
-        ' but NOT update state', () => {
-      resource.state_ = ResourceState_.LAYOUT_COMPLETE;
-      elementMock.expects('unlayoutCallback').returns(false).once();
-      elementMock.expects('togglePlaceholder').withArgs(true).never();
-      resource.unlayout();
-      expect(resource.getState()).to.equal(ResourceState_.LAYOUT_COMPLETE);
-    });
-
-    it('should NOT call viewportCallback when resource not in viewport', () => {
-      resource.state_ = ResourceState_.LAYOUT_COMPLETE;
-      resource.isInViewport_ = false;
-      elementMock.expects('viewportCallback').never();
-      resource.unlayout();
-    });
-
-    it('should call viewportCallback when resource in viewport', () => {
-      resource.state_ = ResourceState_.LAYOUT_COMPLETE;
-      resource.isInViewport_ = true;
-      elementMock.expects('viewportCallback').withExactArgs(false).once();
-      resource.unlayout();
-    });
-
-    it('should delegate unload to unlayoutCallback', () => {
-      resource.state_ = ResourceState_.LAYOUT_COMPLETE;
-      elementMock.expects('unlayoutCallback').returns(false).once();
-      elementMock.expects('togglePlaceholder').withArgs(true).never();
-      resource.unload();
-      expect(resource.getState()).to.equal(ResourceState_.LAYOUT_COMPLETE);
-    });
-  });
-
-  describe('pauseCallback', () => {
-    it('should NOT call pauseCallback on unbuilt element', () => {
-      resource.state_ = ResourceState_.NOT_BUILT;
-      elementMock.expects('pauseCallback').never();
-      resource.pause();
-    });
-
-    it('should NOT call pauseCallback on paused element', () => {
-      resource.state_ = ResourceState_.LAYOUT_COMPLETE;
-      resource.paused_ = true;
-      elementMock.expects('pauseCallback').never();
-      resource.pause();
-    });
-
-    it('should call pauseCallback on built element', () => {
-      resource.state_ = ResourceState_.LAYOUT_COMPLETE;
-      elementMock.expects('pauseCallback').once();
-      resource.pause();
-    });
-
-    it('should NOT call unlayoutCallback', () => {
-      resource.state_ = ResourceState_.LAYOUT_COMPLETE;
-      elementMock.expects('pauseCallback').once();
-      elementMock.expects('unlayoutCallback').never();
-      resource.pause();
-    });
-
-    describe('when unlayoutOnPause', () => {
-      beforeEach(() => {
-        elementMock.expects('unlayoutOnPause').returns(true).once();
-      });
-
-      it('should call unlayoutCallback and update state', () => {
-        resource.state_ = ResourceState_.LAYOUT_COMPLETE;
-        elementMock.expects('pauseCallback').once();
-        elementMock.expects('unlayoutCallback').returns(true).once();
-        resource.pause();
-        expect(resource.getState()).to.equal(ResourceState_.NOT_LAID_OUT);
-      });
-
-      it('should call unlayoutCallback but NOT update state', () => {
-        resource.state_ = ResourceState_.LAYOUT_COMPLETE;
-        elementMock.expects('pauseCallback').once();
-        elementMock.expects('unlayoutCallback').returns(false).once();
-        resource.pause();
-        expect(resource.getState()).to.equal(ResourceState_.LAYOUT_COMPLETE);
-      });
-    });
-  });
-
-  describe('resumeCallback', () => {
-    it('should NOT call resumeCallback on unbuilt element', () => {
-      resource.state_ = ResourceState_.NOT_BUILT;
-      elementMock.expects('resumeCallback').never();
-      resource.resume();
-    });
-
-    it('should NOT call resumeCallback on un-paused element', () => {
-      resource.state_ = ResourceState_.LAYOUT_COMPLETE;
-      elementMock.expects('resumeCallback').never();
-      resource.resume();
-    });
-
-    it('should call resumeCallback on built element', () => {
-      resource.state_ = ResourceState_.LAYOUT_COMPLETE;
-      resource.paused_ = true;
-      elementMock.expects('resumeCallback').once();
-      resource.resume();
-    });
-  });
-
-  describe('getResourcesInViewport', () => {
-    let resource1;
-    let resource2;
-
-    beforeEach(() => {
-      resource1 = {
-        hasOwner: () => false,
-        isDisplayed: () => true,
-        isFixed: () => false,
-        prerenderAllowed: () => true,
-        overlaps: () => true,
-      };
-      resource2 = {
-        hasOwner: () => false,
-        isDisplayed: () => true,
-        isFixed: () => false,
-        prerenderAllowed: () => true,
-        overlaps: () => false,
-      };
-      resources.resources_ = [resource1, resource2];
-    });
-
-    it('should return a subset of resources that are currently ' +
-       'in the viewport', () => {
-      expect(resources.get().length).to.equal(2);
-      expect(resources.getResourcesInViewport().length).to.equal(1);
-    });
-
-    it('should not return resources that are not allowed to prerender if ' +
-       'in prerender mode', () => {
-      resource1.prerenderAllowed = () => false;
-      expect(resources.get().length).to.equal(2);
-      expect(resources.getResourcesInViewport(false).length).to.equal(1);
-      expect(resources.getResourcesInViewport(true).length).to.equal(0);
-    });
-  });
-});
-
-describe('Resource renderOutsideViewport', () => {
-  let sandbox;
-  let element;
-  let elementMock;
-  let resources;
-  let resource;
-  let viewport;
-
-  beforeEach(() => {
-    sandbox = sinon.sandbox.create();
-
-    element = {
-      tagName: 'AMP-AD',
-      isBuilt: () => false,
-      isUpgraded: () => false,
-      prerenderAllowed: () => false,
-      renderOutsideViewport: () => true,
-      build: () => false,
-      getBoundingClientRect: () => null,
-      updateLayoutBox: () => {},
-      isRelayoutNeeded: () => false,
-      layoutCallback: () => {},
-      changeSize: () => {},
-      unlayoutOnPause: () => false,
-      unlayoutCallback: () => true,
-      pauseCallback: () => false,
-      resumeCallback: () => false,
-      viewportCallback: () => {},
-      getPriority: () => 0,
-    };
-    elementMock = sandbox.mock(element);
-
-    resources = new Resources(window);
-    resource = new Resource(1, element, resources);
-    viewport = resources.viewport_;
-    sandbox.stub(viewport, 'getRect').returns(layoutRectLtwh(0, 0, 100, 100));
-  });
-
-  afterEach(() => {
-    elementMock.verify();
-    sandbox.restore();
-  });
-
-
-  describe('boolean API', () => {
-    describe('when element returns true', () => {
-      beforeEach(() => {
-        elementMock.expects('renderOutsideViewport').returns(true).once();
-      });
-
-      describe('when element is inside viewport', () => {
-        it('should allow rendering when bottom falls outside', () => {
-          resource.layoutBox_ = layoutRectLtwh(0, 10, 100, 100);
-          expect(resource.renderOutsideViewport()).to.equal(true);
-        });
-
-        it('should allow rendering when top falls outside', () => {
-          resource.layoutBox_ = layoutRectLtwh(0, -10, 100, 100);
-          expect(resource.renderOutsideViewport()).to.equal(true);
-        });
-      });
-
-      describe('when element is just below viewport', () => {
-        beforeEach(() => {
-          resource.layoutBox_ = layoutRectLtwh(0, 110, 100, 100);
-        });
-
-        it('should allow rendering when scrolling towards', () => {
-          resources.lastVelocity_ = 2;
-          expect(resource.renderOutsideViewport()).to.equal(true);
-        });
-
-        it('should allow rendering when scrolling away', () => {
-          resources.lastVelocity_ = -2;
-          expect(resource.renderOutsideViewport()).to.equal(true);
-        });
-      });
-
-      describe('when element is marginally below viewport', () => {
-        beforeEach(() => {
-          resource.layoutBox_ = layoutRectLtwh(0, 250, 100, 100);
-        });
-
-        it('should allow rendering when scrolling towards', () => {
-          resources.lastVelocity_ = 2;
-          expect(resource.renderOutsideViewport()).to.equal(true);
-        });
-
-        it('should allow rendering when scrolling away', () => {
-          resources.lastVelocity_ = -2;
-          expect(resource.renderOutsideViewport()).to.equal(true);
-        });
-      });
-
-      describe('when element is wayyy below viewport', () => {
-        beforeEach(() => {
-          resource.layoutBox_ = layoutRectLtwh(0, 1000, 100, 100);
-        });
-
-        it('should allow rendering', () => {
-          expect(resource.renderOutsideViewport()).to.equal(true);
-        });
-
-        it('should allow rendering when scrolling towards', () => {
-          resources.lastVelocity_ = 2;
-          expect(resource.renderOutsideViewport()).to.equal(true);
-        });
-
-        it('should allow rendering when scrolling away', () => {
-          resources.lastVelocity_ = -2;
-          expect(resource.renderOutsideViewport()).to.equal(true);
-        });
-      });
-
-      describe('when element is just above viewport', () => {
-        beforeEach(() => {
-          resource.layoutBox_ = layoutRectLtwh(0, -10, 100, 100);
-        });
-
-        it('should allow rendering when scrolling towards', () => {
-          resources.lastVelocity_ = -2;
-          expect(resource.renderOutsideViewport()).to.equal(true);
-        });
-
-        it('should allow rendering when scrolling away', () => {
-          resources.lastVelocity_ = 2;
-          expect(resource.renderOutsideViewport()).to.equal(true);
-        });
-      });
-
-      describe('when element is marginally above viewport', () => {
-        beforeEach(() => {
-          resource.layoutBox_ = layoutRectLtwh(0, -250, 100, 100);
-        });
-
-        it('should allow rendering when scrolling towards', () => {
-          resources.lastVelocity_ = -2;
-          expect(resource.renderOutsideViewport()).to.equal(true);
-        });
-
-        it('should allow rendering when scrolling away', () => {
-          resources.lastVelocity_ = 2;
-          expect(resource.renderOutsideViewport()).to.equal(true);
-        });
-      });
-
-      describe('when element is wayyy above viewport', () => {
-        beforeEach(() => {
-          resource.layoutBox_ = layoutRectLtwh(0, -1000, 100, 100);
-        });
-
-        it('should allow rendering', () => {
-          expect(resource.renderOutsideViewport()).to.equal(true);
-        });
-
-        it('should allow rendering when scrolling towards', () => {
-          resources.lastVelocity_ = -2;
-          expect(resource.renderOutsideViewport()).to.equal(true);
-        });
-
-        it('should allow rendering when scrolling away', () => {
-          resources.lastVelocity_ = 2;
-          expect(resource.renderOutsideViewport()).to.equal(true);
-        });
-      });
-    });
-
-    describe('when element returns false', () => {
-      beforeEach(() => {
-        elementMock.expects('renderOutsideViewport').returns(false).once();
-      });
-
-      describe('when element is inside viewport', () => {
-        it('should allow rendering when bottom falls outside', () => {
-          resource.layoutBox_ = layoutRectLtwh(0, 10, 100, 100);
-          expect(resource.renderOutsideViewport()).to.equal(false);
-        });
-
-        it('should allow rendering when top falls outside', () => {
-          resource.layoutBox_ = layoutRectLtwh(0, -10, 100, 100);
-          expect(resource.renderOutsideViewport()).to.equal(false);
-        });
-      });
-
-      describe('when element is just below viewport', () => {
-        beforeEach(() => {
-          resource.layoutBox_ = layoutRectLtwh(0, 110, 100, 100);
-        });
-
-        it('should disallow rendering when scrolling towards', () => {
-          resources.lastVelocity_ = 2;
-          expect(resource.renderOutsideViewport()).to.equal(false);
-        });
-
-        it('should disallow rendering when scrolling away', () => {
-          resources.lastVelocity_ = -2;
-          expect(resource.renderOutsideViewport()).to.equal(false);
-        });
-      });
-
-      describe('when element is marginally below viewport', () => {
-        beforeEach(() => {
-          resource.layoutBox_ = layoutRectLtwh(0, 250, 100, 100);
-        });
-
-        it('should disallow rendering when scrolling towards', () => {
-          resources.lastVelocity_ = 2;
-          expect(resource.renderOutsideViewport()).to.equal(false);
-        });
-
-        it('should disallow rendering when scrolling away', () => {
-          resources.lastVelocity_ = -2;
-          expect(resource.renderOutsideViewport()).to.equal(false);
-        });
-      });
-
-      describe('when element is wayyy below viewport', () => {
-        beforeEach(() => {
-          resource.layoutBox_ = layoutRectLtwh(0, 1000, 100, 100);
-        });
-
-        it('should disallow rendering', () => {
-          expect(resource.renderOutsideViewport()).to.equal(false);
-        });
-
-        it('should disallow rendering when scrolling towards', () => {
-          resources.lastVelocity_ = 2;
-          expect(resource.renderOutsideViewport()).to.equal(false);
-        });
-
-        it('should disallow rendering when scrolling away', () => {
-          resources.lastVelocity_ = -2;
-          expect(resource.renderOutsideViewport()).to.equal(false);
-        });
-      });
-
-      describe('when element is just above viewport', () => {
-        beforeEach(() => {
-          resource.layoutBox_ = layoutRectLtwh(0, -10, 100, 100);
-        });
-
-        it('should disallow rendering when scrolling towards', () => {
-          resources.lastVelocity_ = -2;
-          expect(resource.renderOutsideViewport()).to.equal(false);
-        });
-
-        it('should disallow rendering when scrolling away', () => {
-          resources.lastVelocity_ = 2;
-          expect(resource.renderOutsideViewport()).to.equal(false);
-        });
-      });
-
-      describe('when element is marginally above viewport', () => {
-        beforeEach(() => {
-          resource.layoutBox_ = layoutRectLtwh(0, -250, 100, 100);
-        });
-
-        it('should disallow rendering when scrolling towards', () => {
-          resources.lastVelocity_ = -2;
-          expect(resource.renderOutsideViewport()).to.equal(false);
-        });
-
-        it('should disallow rendering when scrolling away', () => {
-          resources.lastVelocity_ = 2;
-          expect(resource.renderOutsideViewport()).to.equal(false);
-        });
-      });
-
-      describe('when element is wayyy above viewport', () => {
-        beforeEach(() => {
-          resource.layoutBox_ = layoutRectLtwh(0, -1000, 100, 100);
-        });
-
-        it('should disallow rendering', () => {
-          expect(resource.renderOutsideViewport()).to.equal(false);
-        });
-
-        it('should disallow rendering when scrolling towards', () => {
-          resources.lastVelocity_ = -2;
-          expect(resource.renderOutsideViewport()).to.equal(false);
-        });
-
-        it('should disallow rendering when scrolling away', () => {
-          resources.lastVelocity_ = 2;
-          expect(resource.renderOutsideViewport()).to.equal(false);
-        });
-      });
-    });
-  });
-
-  describe('number API', () => {
-    beforeEach(() => {
-      elementMock.expects('renderOutsideViewport').returns(3).once();
-    });
-
-    describe('when element is inside viewport', () => {
-      it('should allow rendering when bottom falls outside', () => {
-        resource.layoutBox_ = layoutRectLtwh(0, 10, 100, 100);
-        expect(resource.renderOutsideViewport()).to.equal(true);
-      });
-
-      it('should allow rendering when top falls outside', () => {
-        resource.layoutBox_ = layoutRectLtwh(0, -10, 100, 100);
-        expect(resource.renderOutsideViewport()).to.equal(true);
-      });
-    });
-
-    describe('when element is just below viewport', () => {
-      beforeEach(() => {
-        resource.layoutBox_ = layoutRectLtwh(0, 110, 100, 100);
-      });
-
-      it('should allow rendering when scrolling towards', () => {
-        resources.lastVelocity_ = 2;
-        expect(resource.renderOutsideViewport()).to.equal(true);
-      });
-
-      it('should allow rendering when scrolling away', () => {
-        resources.lastVelocity_ = -2;
-        expect(resource.renderOutsideViewport()).to.equal(true);
-      });
-    });
-
-    describe('when element is marginally below viewport', () => {
-      beforeEach(() => {
-        resource.layoutBox_ = layoutRectLtwh(0, 250, 100, 100);
-      });
-
-      it('should allow rendering when scrolling towards', () => {
-        resources.lastVelocity_ = 2;
-        expect(resource.renderOutsideViewport()).to.equal(true);
-      });
-
-      it('should disallow rendering when scrolling away', () => {
-        resources.lastVelocity_ = -2;
-        expect(resource.renderOutsideViewport()).to.equal(false);
-      });
-    });
-
-    describe('when element is wayyy below viewport', () => {
-      beforeEach(() => {
-        resource.layoutBox_ = layoutRectLtwh(0, 1000, 100, 100);
-      });
-
-      it('should disallow rendering', () => {
-        expect(resource.renderOutsideViewport()).to.equal(false);
-      });
-
-      it('should disallow rendering when scrolling towards', () => {
-        resources.lastVelocity_ = 2;
-        expect(resource.renderOutsideViewport()).to.equal(false);
-      });
-
-      it('should disallow rendering when scrolling away', () => {
-        resources.lastVelocity_ = -2;
-        expect(resource.renderOutsideViewport()).to.equal(false);
-      });
-    });
-
-    describe('when element is just above viewport', () => {
-      beforeEach(() => {
-        resource.layoutBox_ = layoutRectLtwh(0, -10, 100, 100);
-      });
-
-      it('should allow rendering when scrolling towards', () => {
-        resources.lastVelocity_ = -2;
-        expect(resource.renderOutsideViewport()).to.equal(true);
-      });
-
-      it('should allow rendering when scrolling away', () => {
-        resources.lastVelocity_ = 2;
-        expect(resource.renderOutsideViewport()).to.equal(true);
-      });
-    });
-
-    describe('when element is marginally above viewport', () => {
-      beforeEach(() => {
-        resource.layoutBox_ = layoutRectLtwh(0, -250, 100, 100);
-      });
-
-      it('should allow rendering when scrolling towards', () => {
-        resources.lastVelocity_ = -2;
-        expect(resource.renderOutsideViewport()).to.equal(true);
-      });
-
-      it('should disallow rendering when scrolling away', () => {
-        resources.lastVelocity_ = 2;
-        expect(resource.renderOutsideViewport()).to.equal(false);
-      });
-    });
-
-    describe('when element is wayyy above viewport', () => {
-      beforeEach(() => {
-        resource.layoutBox_ = layoutRectLtwh(0, -1000, 100, 100);
-      });
-
-      it('should disallow rendering', () => {
-        expect(resource.renderOutsideViewport()).to.equal(false);
-      });
-
-      it('should disallow rendering when scrolling towards', () => {
-        resources.lastVelocity_ = -2;
-        expect(resource.renderOutsideViewport()).to.equal(false);
-      });
-
-      it('should disallow rendering when scrolling away', () => {
-        resources.lastVelocity_ = 2;
-        expect(resource.renderOutsideViewport()).to.equal(false);
-      });
-    });
-  });
-});
-
-describe('Resources fix IE matchMedia', () => {
-  let sandbox;
-  let clock;
-  let windowApi, windowMock;
-  let platformMock;
-  let resources;
-  let devErrorStub;
-  let schedulePassStub;
-
-  beforeEach(() => {
-    sandbox = sinon.sandbox.create();
-    clock = sandbox.useFakeTimers();
-    resources = new Resources(window);
-    resources.relayoutAll_ = false;
-    resources.doPass_ = () => {};
-    platformMock = sandbox.mock(resources.platform_);
-    devErrorStub = sandbox.stub(dev, 'error');
-    schedulePassStub = sandbox.stub(resources, 'schedulePass');
-
-    windowApi = {
-      innerWidth: 320,
-      setInterval: () => {},
-      clearInterval: () => {},
-      matchMedia: () => {},
-    };
-    windowMock = sandbox.mock(windowApi);
-  });
-
-  afterEach(() => {
-    platformMock.verify();
-    windowMock.verify();
-    sandbox.restore();
-  });
-
-  it('should bypass polling for non-IE browsers', () => {
-    platformMock.expects('isIe').returns(false);
-    windowMock.expects('matchMedia').never();
-    windowMock.expects('setInterval').never();
-    resources.fixMediaIe_(windowApi);
-    expect(resources.relayoutAll_).to.be.true;
-    expect(devErrorStub.callCount).to.equal(0);
-    expect(schedulePassStub.callCount).to.equal(0);
-  });
-
-  it('should bypass polling when matchMedia is not broken', () => {
-    platformMock.expects('isIe').returns(true);
-    windowMock.expects('matchMedia')
-        .withExactArgs('(min-width: 320px) AND (max-width: 320px)')
-        .returns({matches: true})
-        .once();
-    windowMock.expects('setInterval').never();
-    resources.fixMediaIe_(windowApi);
-    expect(resources.relayoutAll_).to.be.true;
-    expect(devErrorStub.callCount).to.equal(0);
-    expect(schedulePassStub.callCount).to.equal(0);
-  });
-
-  it('should poll when matchMedia is wrong, but eventually succeeds', () => {
-    platformMock.expects('isIe').returns(true);
-
-    // Scheduling pass.
-    windowMock.expects('matchMedia')
-        .withExactArgs('(min-width: 320px) AND (max-width: 320px)')
-        .returns({matches: false})
-        .once();
-    const intervalId = 111;
-    let intervalCallback;
-    windowMock.expects('setInterval')
-        .withExactArgs(
-            sinon.match(arg => {
-              intervalCallback = arg;
-              return true;
-            }),
-            10
-        )
-        .returns(intervalId)
-        .once();
-
-    resources.fixMediaIe_(windowApi);
-    expect(resources.relayoutAll_).to.be.false;
-    expect(devErrorStub.callCount).to.equal(0);
-    expect(schedulePassStub.callCount).to.equal(0);
-    expect(intervalCallback).to.exist;
-    windowMock.verify();
-    windowMock./*OK*/restore();
-
-    // Second pass.
-    clock.tick(10);
-    windowMock = sandbox.mock(windowApi);
-    windowMock.expects('matchMedia')
-        .withExactArgs('(min-width: 320px) AND (max-width: 320px)')
-        .returns({matches: false})
-        .once();
-    windowMock.expects('clearInterval').never();
-    intervalCallback();
-    expect(resources.relayoutAll_).to.be.false;
-    expect(devErrorStub.callCount).to.equal(0);
-    expect(schedulePassStub.callCount).to.equal(0);
-    windowMock.verify();
-    windowMock./*OK*/restore();
-
-    // Third pass - succeed.
-    clock.tick(10);
-    windowMock = sandbox.mock(windowApi);
-    windowMock.expects('matchMedia')
-        .withExactArgs('(min-width: 320px) AND (max-width: 320px)')
-        .returns({matches: true})
-        .once();
-    windowMock.expects('clearInterval').withExactArgs(intervalId).once();
-    intervalCallback();
-    expect(resources.relayoutAll_).to.be.true;
-    expect(schedulePassStub.callCount).to.equal(1);
-    expect(devErrorStub.callCount).to.equal(0);
-    windowMock.verify();
-    windowMock./*OK*/restore();
-  });
-
-  it('should poll until times out', () => {
-    platformMock.expects('isIe').returns(true);
-
-    // Scheduling pass.
-    windowMock.expects('matchMedia')
-        .withExactArgs('(min-width: 320px) AND (max-width: 320px)')
-        .returns({matches: false})
-        .atLeast(2);
-    const intervalId = 111;
-    let intervalCallback;
-    windowMock.expects('setInterval')
-        .withExactArgs(
-            sinon.match(arg => {
-              intervalCallback = arg;
-              return true;
-            }),
-            10
-        )
-        .returns(intervalId)
-        .once();
-    windowMock.expects('clearInterval').withExactArgs(intervalId).once();
-
-    resources.fixMediaIe_(windowApi);
-    expect(resources.relayoutAll_).to.be.false;
-    expect(devErrorStub.callCount).to.equal(0);
-    expect(schedulePassStub.callCount).to.equal(0);
-    expect(intervalCallback).to.exist;
-
-    // Second pass.
-    clock.tick(10);
-    intervalCallback();
-    expect(resources.relayoutAll_).to.be.false;
-    expect(devErrorStub.callCount).to.equal(0);
-    expect(schedulePassStub.callCount).to.equal(0);
-
-    // Third pass - timeout.
-    clock.tick(2000);
-    intervalCallback();
-    expect(resources.relayoutAll_).to.be.true;
-    expect(schedulePassStub.callCount).to.equal(1);
-    expect(devErrorStub.callCount).to.equal(1);
-  });
-
-  it('should tolerate matchMedia exceptions', () => {
-    platformMock.expects('isIe').returns(true);
-
-    windowMock.expects('matchMedia')
-        .withExactArgs('(min-width: 320px) AND (max-width: 320px)')
-        .throws(new Error('intentional'))
-        .once();
-    windowMock.expects('setInterval').never();
-
-    resources.fixMediaIe_(windowApi);
-    expect(resources.relayoutAll_).to.be.true;
-    expect(devErrorStub.callCount).to.equal(1);
-    expect(schedulePassStub.callCount).to.equal(0);
-  });
-});
-
-
 describe('Resources.add', () => {
   let sandbox;
   let resources;
@@ -2632,7 +1342,7 @@ describe('Resources.add', () => {
   function createElementWithResource(id) {
     const element = createElement();
     const resource = new Resource(id, element, resources);
-    resource.state_ = ResourceState_.NOT_BUILT;
+    resource.state_ = ResourceState.NOT_BUILT;
     resource.element['__AMP__RESOURCE'] = resource;
     return [element, resource];
   }
