+++ test/functional/test-task-queue.js
@@ -1,0 +1,68 @@
+/**
+ * Copyright 2016 The AMP HTML Authors. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS-IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import {TaskQueue} from '../../src/service/task-queue';
+import * as sinon from 'sinon';
+
+
+describe('TaskQueue', () => {
+
+  let sandbox;
+  let clock;
+  let queue;
+
+  beforeEach(() => {
+    sandbox = sinon.sandbox.create();
+    clock = sandbox.useFakeTimers();
+    queue = new TaskQueue();
+  });
+
+  afterEach(() => {
+    sandbox.restore();
+  });
+
+  it('should enqueue and dequeue', () => {
+    clock.tick(1000);
+    expect(queue.getSize()).to.equal(0);
+    expect(queue.getLastEnqueueTime()).to.equal(0);
+    expect(queue.getLastDequeueTime()).to.equal(0);
+
+    queue.enqueue({id: '1'});
+    expect(queue.getTaskById('1').id).to.equal('1');
+    expect(queue.getSize()).to.equal(1);
+    expect(queue.getLastEnqueueTime()).to.equal(1000);
+    expect(queue.getLastDequeueTime()).to.equal(0);
+
+    expect(() => {
+      queue.enqueue({id: '1'});
+    }).to.throw(/Task already enqueued/);
+
+    queue.dequeue({id: '1'});
+    expect(queue.getTaskById('1')).to.equal(null);
+    expect(queue.getSize()).to.equal(0);
+    expect(queue.getLastEnqueueTime()).to.equal(1000);
+    expect(queue.getLastDequeueTime()).to.equal(1000);
+  });
+
+  it('should perform score-based peek', () => {
+    queue.enqueue({id: 'A', v: 0});
+    queue.enqueue({id: 'B', v: 2});
+    queue.enqueue({id: 'C', v: 1});
+
+    const task = queue.peek(task => 10 - task.v);
+    expect(task.id).to.equal('B');
+  });
+});
